refactor(agent): extract composer history hook
